third_party/WebKit/LayoutTests/external/wpt/webvtt/webvtt-api-for-browsers/vttcue-interface/line-expected.txt

CONSOLE ERROR: Text track from origin 'null' has been blocked from loading: Not at same origin as the document, and parent of track element does not have a 'crossorigin' attribute. Origin 'http://web-platform.test:8001' is therefore not allowed access.
This is a testharness.js-based test.
PASS VTTCue.line, script-created cue 
FAIL VTTCue.line, parsed cue assert_unreached: got error event Reached unreachable code
Harness: the test ran to completion.
